utils: Moved Lists to sep file
