+++ src/core/utils.hpp
@@ -41,6 +41,7 @@
 #include "lees_edwards.hpp"
 #include "utils/math/sqr.hpp"
 #include "utils/memory.hpp"
+#include "utils/List.hpp"
 
 /*************************************************************/
 /** \name Mathematical, physical and chemical constants.     */
@@ -77,40 +78,6 @@
  ************************************************/
 
 extern int this_node;
-
-/** Integer list.
-    Use the functions specified in list operations. */
-typedef struct {
-  int *begin() { return e; }
-  int *end() { return e + n; }
-  int size() const { return n; }
-  void resize(int size) {
-    if (size != this->max) {
-      this->max = size;
-      this->e =
-          static_cast<int *>(Utils::realloc(this->e, sizeof(int) * this->max));
-    }
-  }
-  /** Dynamically allocated integer field. */
-  int *e;
-  /** number of used elements in the integer field. */
-  int n;
-  /** allocated size of the integer field. This value is ONLY changed
-      in the routines specified in list operations ! */
-  int max;
-} IntList;
-
-/** Double list.
-    Use the functions specified in list operations. */
-typedef struct {
-  /** Dynamically allocated double field. */
-  double *e;
-  /** number of used elements in the double field. */
-  int n;
-  /** allocated size of the double field. This value is ONLY changed
-      in the routines specified in list operations ! */
-  int max;
-} DoubleList;
 
 namespace Utils {
 /**
